rag-demo/backend/requirements.txt
# Django 核心
Django>=4.0
djangorestframework>=3.14
django-cors-headers>=4.0

# LangChain 核心
langchain>=0.1.0
langchain-community>=0.0.10
langchain-openai>=0.0.5
langchain-huggingface>=0.0.1

# 向量数据库
chromadb>=0.4.0

# OpenAI API
openai>=1.0.0

# 环境变量
python-dotenv>=1.0.0

# Sentence Transformers (用于 HuggingFaceEmbeddings)
sentence-transformers>=2.2.0
